apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-multipart==0.0.6
langchain-openai==0.0.8
celery==5.3.4
redis==5.0.1
loguru==0.7.2
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.7
langchain-core==0.1.0
PyPDF2==3.0.1
python-docx==1.1.0
rank-bm25==0.2.2
jieba==0.42.1